implement new price structure

+++ app/Livewire/Architects/AddStories/PressReleaseDraft.php
@@ -5,6 +5,7 @@ namespace App\Livewire\Architects\AddStories;
 ini_set('max_execution_time', 300);
 
 use App\Http\Controllers\Users\Architects\MediaKitDraftController;
+use App\Http\Controllers\Users\MediaKitController;
 use App\Livewire\Forms\Architects\PressReleaseForm;
 use Livewire\Component;
 use Livewire\Features\SupportFileUploads\TemporaryUploadedFile;
@@ -63,6 +64,10 @@ class PressReleaseDraft extends Component
 
 	public function add()
 	{
+		if(!MediaKitController::isAllowedToAdd('press-release')){
+			$this->dispatch('show-media-kit-limit-alert-modal');
+			return;
+		}
 		if($this->form->store()){
 			MediaKitDraftController::deleteById($this->draftId);
 			$this->dispatch('alert', [
